+++ .devcontainer/devcontainer.json
@@ -27,7 +27,6 @@
                 "github.vscode-github-actions",
                 "GitHub.vscode-pull-request-github",
                 "redhat.vscode-yaml",
-                "vscjava.vscode-maven",
                 "redhat.vscode-xml"
             ],
             "settings": {
