+++ src/java/com/kerosenelabs/kindling/WorkerThreadNameGenerator.java
@@ -1,0 +1,5 @@
+package com.kerosenelabs.kindling;
+
+public interface WorkerThreadNameGenerator {
+    public String generate();
+}
