+++ src/java/com/kerosenelabs/kindling/constant/HttpMethod.java
@@ -1,0 +1,8 @@
+package com.kerosenelabs.kindling.constant;
+
+public enum HttpMethod {
+    GET,
+    PUT,
+    POST,
+    DELETE,
+}
