+++ src/java/com/kerosenelabs/kindling/constant/MimeType.java
@@ -1,0 +1,23 @@
+package com.kerosenelabs.kindling.constant;
+
+public enum MimeType {
+    APPLICATION_JSON("application/json"),
+    APPLICATION_XML("application/xml"),
+    TEXT_PLAIN("text/plain"),
+    TEXT_HTML("text/html");
+
+    private final String mimeType;
+
+    MimeType(String mimeType) {
+        this.mimeType = mimeType;
+    }
+
+    public String getMimeType() {
+        return mimeType;
+    }
+
+    @Override
+    public String toString() {
+        return mimeType;
+    }
+}
