--- src/main/java/io/kerosenelabs/kindling/WorkerThreadNameGenerator.java
@@ -1,5 +1,0 @@
-package io.kerosenelabs.kindling;
-
-public interface WorkerThreadNameGenerator {
-    public String generate();
-}
