--- src/main/java/io/kerosenelabs/kindling/constant/HttpMethod.java
@@ -1,8 +1,0 @@
-package io.kerosenelabs.kindling.constant;
-
-public enum HttpMethod {
-    GET,
-    PUT,
-    POST,
-    DELETE,
-}
