--- src/main/java/io/kerosenelabs/kindling/exception/KindlingException.java
@@ -1,11 +1,0 @@
-package io.kerosenelabs.kindling.exception;
-
-public class KindlingException extends Exception {
-    public KindlingException(Throwable e) {
-        super(e);
-    }
-
-    public KindlingException(String message) {
-        super(message);
-    }
-}
