+++ src/test/java/io/kerosenelabs/kindling/HttpServerTest.java
@@ -1,5 +1,5 @@
-package io.kerosenelabs.kindling;
+package com.kerosenelabs.kindling;
 
 public class HttpServerTest {
-    
+
 }
